this has been great, we got it to work
